+++ components/editor/EditablePage/index.tsx
@@ -5,7 +5,7 @@ import cx from 'classnames'
 import ReactDragListView from 'react-drag-listview'
 
 import definedBlocks from 'lib/blocks.json'
-import { EditableBlock } from 'components/_editor/EditableBlock'
+import { EditableBlock } from 'components/editor/EditableBlock'
 import { titleAtom } from 'lib/atoms/form'
 import { formBlock, formStyle } from 'lib/types/form'
 import { placeCaretAtEnd } from 'lib/utils/placeCaretAtEnd'
